+++ CHANGELOG.md
@@ -55,6 +55,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More complete simplification during interpretation
 - SMT-based resolving of addresses now works for delegatecall and staticcall
   opcodes as well
+- UNSAT cache is now in `Solvers.hs` and is therefore shared across all threads.
+  Hence, it is now active even during branch queries.
 
 ## Fixed
 - We now try to simplify expressions fully before trying to cast them to a concrete value